requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.6
PyJWT==1.7.1
pylint==2.6.0
python-decouple==3.3
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
twilio==6.49.0
urllib3==1.26.2
virtualenv==20.2.1
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.12.1
